feat: make Square directory search-ready
