+++ web/src/app/oauth/layout.tsx
@@ -1,0 +1,12 @@
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false
+  }
+};
+
+export default function OAuthLayout({ children }: { children: React.ReactNode }) {
+  return children;
+}
