+++ web/src/app/robots.ts
@@ -1,0 +1,12 @@
+import type { MetadataRoute } from 'next';
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: '*',
+      allow: '/',
+      disallow: ['/add', '/oauth', '/notification', '/setting']
+    },
+    sitemap: ['https://square.degov.ai/sitemap.xml']
+  };
+}
